Don't use npm list --parseable
Turns out we lose version info with --parseable (so not sure in what
world this is parseable output) so stop using the --parseable flag. This
corrects behavior where we reinstall ethercalc every puppet run which
can upgrade unpinned dep libs in ways that break them on older node (we
are using the ethercalc prescribed node but its a little old).

Change-Id: Ib9aebecf3a0c36c50e5069b1a871886e4ddb1a42